refactor: move cli session modules into cli/session/ subpackage
Create session/ with _cook, _order, _session_launch, _session_picker,
_reload. Update all intra-session and cross-subpackage references.
