+++ src/autoskillit/cli/fleet/_fleet_session.py
@@ -38,7 +38,7 @@ def _launch_fleet_session(
 ) -> None:
     """Build the L3 orchestrator prompt and launch an interactive fleet session."""
     from autoskillit.cli import detect_autoskillit_mcp_prefix  # noqa: PLC0415
-    from autoskillit.cli._session_launch import _run_interactive_session
+    from autoskillit.cli.session._session_launch import _run_interactive_session
 
     mcp_prefix = detect_autoskillit_mcp_prefix()
