+++ src/autoskillit/cli/session/__init__.py
@@ -1,0 +1,4 @@
+"""CLI session management — cook, order, resume, reload."""
+
+from autoskillit.cli.session._cook import cook  # noqa: F401
+from autoskillit.cli.session._order import order  # noqa: F401
