+++ src/autoskillit/cli/session/_session_picker.py
